// lib/enums.ts
export enum QuestionContentType {
  Text = 'text',
  Voice = 'voice',
  Meaning = 'meaning',
  Blank = 'blank',
}

export enum AnswerContentType {
  Input = 'input',
  Choice = 'choice',
}

export enum Language {
  Vietnamese = 'vi',
  English = 'en',
}

export enum PartOfSpeech {
  Noun = 'noun',
  Verb = 'verb',
  Adjective = 'adjective',
  Adverb = 'adverb',
  Pronoun = 'pronoun',
  Preposition = 'preposition',
  Conjunction = 'conjunction',
  Interjection = 'interjection',
}